[feature] add working tests for faq on chrome and firefox
