+++ src/test/java/BaseTest.java
@@ -1,0 +1,31 @@
+import org.junit.After;
+import org.junit.Before;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.firefox.FirefoxOptions;
+
+public class BaseTest {
+    public WebDriver driver;
+
+    @Before
+    public void setUp(){
+//        ChromeOptions options = new ChromeOptions();
+//        options.addArguments("--no-sandbox", "--disable-dev-shm-usage", "--remote-allow-origins=*");
+//        System.setProperty("webdriver.chrome.driver", "C:\\Users\\k.nistratov\\chromedriver.exe");
+//        driver = new ChromeDriver(options);
+
+        FirefoxOptions options = new FirefoxOptions();
+        options.addArguments("--no-sandbox", "--disable-dev-shm-usage", "--remote-allow-origins=*");
+        driver = new FirefoxDriver();
+
+
+        driver.get("https://qa-scooter.praktikum-services.ru/");
+    }
+
+    @After
+    public void tearDown(){
+        driver.quit();
+    }
+}
