--- src/test/java/firefox/FaqSectionTest.java
@@ -1,12 +1,0 @@
-package firefox;
-
-import org.junit.Test;
-import pages.FaqPageModel;
-
-public class FaqSectionTest extends BaseFirefoxTest {
-
-    @Test
-    public void openQuestionAndCheck(){
-        new FaqPageModel(driver).openAllArrowsAndCheck();
-    }
-}
